# trusted-pool-platform/api/openapi.yaml
openapi: 3.1.0
info:
  title: 可信有限成员订阅共享平台 Phase 2-A API
  version: 0.2.0
  description: |
    与当前 Go HTTP 实现一致的 Phase 2-A 持久运行时接口。Provision、Seat/Owner Assignment、Operation
    和 Provision credential claim 使用 PostgreSQL；数据库或 KMS 不可用时失败关闭。暂停、换员、结算、
    credential batch 和 control rotation evidence 尚未接入持久工作流，当前统一返回 503。
    风险窗口仍是可丢失的内存观察信号，不自动暂停。普通接口使用服务 Bearer API Key。
...
servers:
  - url: http://localhost:8092
security:
  - serviceBearer: []
paths:
...
  /api/v1/seats:
    post:
      operationId: provisionSeat
      summary: 在既有 Sub2API Group 中开通 Seat
      description: |
        平台先以 operation_id 幂等调用 Sub2API 创建不可交互 Principal、Subscription 和 API Key，
        仅在上游明确返回 ACTIVE 后创建本地 Seat。结果未知时返回 503，调用方必须以完全相同的请求体和
        operation_id 重放；成功响应不直接包含 credential。claim token 仅在首次明确成功响应中返回，
        平台只保存其 SHA-256 摘要，因此后续 provision 重放不会再次返回 token。
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProvisionSeatRequest'}
      responses:
        '201':
          description: Sub2API 已明确完成开通，平台本地 Seat 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvisionSeatEnvelope'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503':
          description: 上游不可用或结果未知；不得推断 Seat 已创建，须使用同一 operation_id 重放
...
  /api/v1/seats/{id}/settlements:
    get:
      operationId: listPendingSettlements
      summary: 列出阻断 Seat 冻结的待对账结算
      description: Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 100}
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/seats/{id}/settlements/{settlement_id}:
    get:
      operationId: getPendingSettlement
      summary: 查询单条待对账结算
      description: Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503。
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/SettlementId'
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/seats/{id}/settlements/{settlement_id}/resolve:
    post:
      operationId: resolvePendingSettlement
      summary: 人工核验后解除单条结算屏障
      description: |
        Phase 2-A PostgreSQL Runtime 尚未接线，当前返回 503。

        必须使用独立 TRUSTED_POOL_SETTLEMENT_API_KEY。平台把 operation_id 同时作为 Sub2API
        Idempotency-Key 提交；Sub2API 继续使用独立 settlement:resolve scope 校验集成客户端并写审计。
      security:
        - settlementBearer: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/SettlementId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResolvePendingSettlementRequest'}
      responses:
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
components:
  securitySchemes:
    serviceBearer:
      type: http
      scheme: bearer
      description: TRUSTED_POOL_API_KEY；仅供受控服务调用，不是最终用户会话或 RBAC
    settlementBearer:
      type: http
      scheme: bearer
      description: TRUSTED_POOL_SETTLEMENT_API_KEY；仅用于人工解除 pending settlement
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
    SettlementId:
      name: settlement_id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
  schemas:
...
    ProvisionSeatRequest:
      type: object
      additionalProperties: false
      required: [id, pool_id, owner_user_id, operation_id, existing_group_id, subscription_expires_at]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        pool_id: {type: string, minLength: 1, maxLength: 128}
        owner_user_id: {type: string, minLength: 1, maxLength: 128}
        operation_id: {type: string, minLength: 1, maxLength: 128}
        existing_group_id:
          type: integer
          format: int64
          minimum: 1
          description: 管理员预先准备、尚未绑定或已绑定同一 pool_id 的 Sub2API 独占 Group ID
        subscription_expires_at: {type: string, format: date-time}
        principal_concurrency:
          type: integer
          minimum: 1
          default: 1
        principal_rpm_limit:
          type: integer
          minimum: 0
          default: 0
        api_key_quota: {type: number, minimum: 0, default: 0}
        api_key_rate_limit_5h: {type: number, minimum: 0, default: 0}
        api_key_rate_limit_1d: {type: number, minimum: 0, default: 0}
        api_key_rate_limit_7d: {type: number, minimum: 0, default: 0}
    ProvisionSeatEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [seat, operation]
          properties:
            seat: {$ref: '#/components/schemas/Seat'}
            operation: {$ref: '#/components/schemas/Operation'}
    Seat:
      type: object
      required: [id, pool_id, state, owner_user_id, current_assignment, assignment_epoch, membership_epoch, updated_at]
      properties:
        id: {type: string}
        pool_id: {type: string}
        state:
          type: string
          enum: [ACTIVE, SUSPEND_PENDING, DRAINING, FROZEN, ASSIGNMENT_PENDING, REPLACEMENT_PENDING]
        owner_user_id: {type: string}
        current_assignment: {$ref: '#/components/schemas/Assignment'}
        assignment_epoch: {type: integer, format: int64, minimum: 1}
        membership_epoch: {type: integer, format: int64, minimum: 1}
        freeze_snapshot: {$ref: '#/components/schemas/FreezeSnapshot'}
        pending_change:
          type: object
          additionalProperties: true
        updated_at: {type: string, format: date-time}
    Assignment:
      type: object
      required: [user_id, kind, assignment_epoch, assigned_at]
      properties:
        user_id: {type: string}
        kind: {type: string, enum: [OWNER, TEMPORARY]}
        assignment_epoch: {type: integer, format: int64}
        assigned_at: {type: string, format: date-time}
    FreezeSnapshot:
      type: object
      required: [operation_id, assignment_epoch, usage, window_starts, in_flight, pending_settlements, captured_at]
      properties:
        operation_id: {type: string}
        assignment_epoch: {type: integer, format: int64}
        usage:
          type: object
          additionalProperties: {type: number}
        window_starts:
          type: object
          description: 上游原生窗口尚未启动时，对应值保持 null
          additionalProperties:
            type: [string, 'null']
            format: date-time
        in_flight: {type: integer, const: 0}
        pending_settlements:
          type: integer
          const: 0
          description: 冻结快照持久化的结算屏障零值证据
        captured_at: {type: string, format: date-time}
...
    Operation:
      type: object
      required: [id, kind, seat_id, status, created_at, updated_at]
      properties:
        id: {type: string}
        kind: {type: string, enum: [PROVISION, SUSPEND, ASSIGN_TEMPORARY, RESTORE, REPLACE_PERMANENTLY]}
        seat_id: {type: string}
        target_user_id: {type: string}
        control_rotation_evidence_ref:
          type: string
          description: 仅永久换员存在；本平台签发并按 Seat Pool 与相邻 Membership Epoch 复核的证据 ID
        status: {type: string, enum: [RUNNING, SUCCEEDED, RETRYABLE, RECONCILE_REQUIRED, FAILED]}
        reason: {type: string}
        credential_claim_token:
          type: string
          x-sensitive: true
          description: 仅随首次开通或换员明确成功响应返回的随机一次性领取令牌；平台不保存明文，操作查询和幂等重放不会返回
        credential_claim_status:
          type: string
          enum: [READY, ACK_PENDING, ACK_RECONCILE_REQUIRED, CLAIMED, EXPIRED, REJECTED]
        credential_claim_expires_at:
          type: string
          format: date-time
          description: 到期后平台主动清除待交付 credential 与 token 摘要
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    ResolvePendingSettlementRequest:
      type: object
      additionalProperties: false
      required: [operation_id, reason, evidence]
      properties:
        operation_id: {type: string, minLength: 1, maxLength: 128}
        reason: {type: string, minLength: 1, maxLength: 1000}
        evidence: {type: string, minLength: 1, maxLength: 4000}
...
    ErrorEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [reason, message]
          properties:
            reason: {type: string}
            message: {type: string}
  responses:
...
    BadRequest:
      description: 请求无效
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Unauthorized:
      description: Bearer 服务凭据缺失或无效
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Conflict:
      description: Seat 状态、冻结快照或幂等 Key 冲突
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    BadGateway:
      description: Sub2API 返回服务端错误或成功响应未通过一致性校验
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    ServiceUnavailable:
      description: 持久工作流尚未接线，或 PostgreSQL/KMS/上游暂时不可用
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}